+++ src/index.ts
@@ -1,6 +1,4 @@
 export * from './comparator';
 export * from './comparisons';
-export * from './config';
-export * from './matching';
-export * from './tracked';
-export * from './utils';
+export * from './utils/tracked';
+export * from './utils/utils';
